[COST-5502] Skip tests if label is set on PR (#5386)
A hard coded timeout of 2 hours is blocking our ability to move fully to running tests in Konflux.

https://issues.redhat.com/browse/KONFLUX-5449

In the mean time, a label can be used to toggle between running tests in Konflux or Jenkins. This prevents consuming unnecessary resources.

The integration test configuration in Konflux remains in place, making the change to fully using Konflux for CI testing a minor code change in koku-test-container once the timeout issue is fixed.